Add Panduan

+++ app/Http/Controllers/VersiPertanyaanController.php
@@ -17,7 +17,7 @@ class VersiPertanyaanController extends Controller
 
         return response()->json([
             'status' => 'success',
-            'message' => __('display_data', ['data' => 'Versi Pertanyaan']),
+            'message' => __('display_data', ['data' => 'versi pertanyaan']),
             'data' => $versiPertanyaan
         ]);
     }
@@ -28,10 +28,13 @@ class VersiPertanyaanController extends Controller
     public function store(StoreVersiPertanyaanRequest $request)
     {
         $validated = $request->validated();
+        if ($validated['status'] === 'Active') {
+            VersiPertanyaan::where('status', 'Active')->update(['status' => 'Inactive']);
+        }
         $versiPertanyaan = VersiPertanyaan::create($validated);
         return response()->json([
             'status' => 'success',
-            'message' => __('create_data', ['data' => 'Versi Pertanyaan']),
+            'message' => __('create_data', ['data' => 'versi pertanyaan']),
             'data' => $versiPertanyaan
         ]);
     }
@@ -43,7 +46,7 @@ class VersiPertanyaanController extends Controller
     {
         return response()->json([
             'status' => 'success',
-            'message' => __('detail_data', ['data' => 'Versi Pertanyaan']),
+            'message' => __('detail_data', ['data' => 'versi pertanyaan']),
             'data' => $versiPertanyaan
         ]);
     }
@@ -54,10 +57,13 @@ class VersiPertanyaanController extends Controller
     public function update(UpdateVersiPertanyaanRequest $request, VersiPertanyaan $versiPertanyaan)
     {
         $validated = $request->validated();
+        if ($validated['status'] === 'Active') {
+            VersiPertanyaan::where('status', 'Active')->update(['status' => 'Inactive']);
+        }
         $versiPertanyaan->update($validated);
         return response()->json([
             'status' => 'success',
-            'message' => __('update_data', ['data' => 'Versi Pertanyaan']),
+            'message' => __('update_data', ['data' => 'versi pertanyaan']),
             'data' => $versiPertanyaan
         ]);
     }
@@ -67,11 +73,17 @@ class VersiPertanyaanController extends Controller
      */
     public function destroy(VersiPertanyaan $versiPertanyaan)
     {
-        $versiPertanyaan->delete();
-        return response()->json([
-            'status' => 'success',
-            'message' => __('delete_data', ['data' => 'Versi Pertanyaan']),
-            'data' => $versiPertanyaan
-        ]);
+        if ($versiPertanyaan->status == 'Active') {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Tidak bisa menghapus versi pertanyaan aktif',
+            ]);
+        } else {
+            $versiPertanyaan->delete();
+            return response()->json([
+                'status' => 'success',
+                'message' => __('delete_data', ['data' => 'versi pertanyaan']),
+            ]);
+        }
     }
 }
